Update beardtown extension
- Lint fixes
- Split out code
